Add missing configuration options for the Ada Language Server.
Additionally, prefer "initialization-options" over "initialized-fn" as
the Ada Language Server prefers to know the project file as part of
initialization, and will have to create a default one if it's not
specified as part of these options.  If a project file is not
specified as part of the initialization options, the server might also
display a diagnostic about no project being specified until a
workspace/didChangeConfiguration is received specifying the project
file.

Due to a current issue with the Ada Language Server, "initialized-fn"
is kept to work around an issue where the project specified in
"initialization-options" might not be honored.

Add support for Ada run-time library folders via the new user option
`lsp-ada-library-folders`, defaulting to detecting the run-time
directory with it's unique "adainclude" folder.