Deal with 'workspace/diagnostic/refresh' message
lsp-mode does not currently support pull diagnostics.

There is a recent change to the spec where the server can notify the
client to request diagnostics, when the pull mode is being used.

This is implemented incorrectly in a number of servers, which do not
first check the relevant client capability.

Mitigate this problem by
a. Sending client capabilities to day that this is not supported
b. Putting a null handler in for the server request message.

Closes #4570