fix: use a transient map to avoid overlay keymaps issues
When some other overlay is active at point with a keymap property, even when
its priority is lower than the inline completion priority, this foreign keymap
may override the inline completion keymap.

This happens, for example, when the inline completion is shown inside a pair
inserted by smartparens. For example, the user types `[` followed `<return>`
-- smartparens will insert the closing pair. The result is the buffer with the
following state (cursor at `|`:

```
list = [
    |
]
```

Smartparens will have placed an overlay from `[` to up `]`. This overlay has a
keymap property mapping `C-g` to a function that removes the overlay.

When inline completion is shown (either on idle or by user request), the
keymap from smartparens overlay is active despite the inline completion
overlay's higher priority.

As a result, pressing `C-<return>` will likely display a message complaining
the key is not bound.

If the user presses `C-g` once, then they gain access to the inline completion
keymap.

This caused a weird bug in which a user gets a suggestion, but can't accept
it. Pressing C-g only once would not cancel the completion, but pressing it
again would indeed high the completion overlay. Explicitly asking for a
suggestion at the same point would then display the completion overlay and the
keymap would work as expected, since the first C-g removed the smartparens
overlay.

This commit fixes the issue by using a transient map when displaying the
overlay. The transient map will take priority over any active overlay map, so
we do not run into this issue from smartparens or any other mode that may be
placing overlays with active keymaps.