OCaml: Add implIntf (#4732)
*This is the first of a series of PR I aim to do to improve the ocaml-lsp
experience in lsp-mode*

This PR exposes `lsp-ocaml-find-alternate-file` that allows to find the file
interface corresponding to an implementation or the implementation corresponding
to an interface.

This function is quite simple for the moment since I first needed to make sure
that the LSP requests were working as expected.

Following PRs will add more possibilities