Add support for Python(ty) (#4785)
* feat(clients): add LSP support for Python(ty) language

Introduce a new client for the Python(ty) programming language in `lsp-mode`.
This includes the ability to start the `ty` language server and provides integration with `python-mode`.

* feat(python): add support for Python(ty) language server

Added configuration and documentation for the Python(ty) language server, including links to its repository and installation instructions.
Updated the changelog and navigation in the documentation to reflect the new support.

* feat(lsp): add support for lsp-python-ty