Make sure lsp-terraform functions called with correct keys
lsp-terraform functions/macros define keys via underscore, but at some
locations the functions were called with dash used inside key names.
This was resulting in the mismatch and a compilation failure.

Correct the call places to use the correct key names.

NB: underscores is not a typical naming convention, but I left it as
is for 2 reasons:

1. I can't be sure if I rename this to dash, compilation wouldn't
break invisibly elsewhere in the mode. The bug being fixed illustrates
that whatever compilation code lsp-mode uses is invalid, because prior
to this commit the file `lsp-terraform.el` would fail to compile if
you try to M-x `byte-compile` it, however `make compile` passes fine.
2. It is an old code, so people might be dependent on the current
naming in their configurations.

Fixes: https://github.com/emacs-lsp/lsp-mode/issues/3577