fix: defer onTypeFormatting to after running other hooks
If other functions on post-self-insert-hook (e.g. smart-parens,
electric-indent) run after LSP's on-type-formatting, they may modify the
buffer and therefore cancel the async reques and prevent on-type-formatting
from working as expected.

This commit updates the hook to run in a timer (after 0 seconds), so it lets
all other functions perform their changes before asking the server for a
formatting action.