fix: cancel inline completion UI on non mapped commands
On some situations, keybindings such as M-x are not caught by the default
handler (lsp-inline-completion-cancel-with-input, via
overriding-terminal-local-map). This results in the tooltip not being hidden
when it should. The consequence is that the user has to explicitly call
lsp-inline-completion-cancel to hide the completion overlay and continue
working in the buffer

This change fixes that by placing a symbol property on the commands that
should not cancel the UI (accept, next, prev, recenter-top-botton and ignore).
On pre-command hook, we check for that property and cancel the completion if
needed.