clients/lsp-trunk.el - prevent error when using lsp on non-file buffer (#4885)
* clients/lsp-trunk.el - prevent error when using lsp on non-file buffer

Protect against error when (file-name-directory filename) is nil which occurs on non-file buffer. We
were hitting this when using lsp-mode with the p4.el package which when diff'ing pulls files from
Perforce into non-file buffers and then activates a major mode that runs lsp.  This ends up calling
lsp-trunk-check-for-init.

* clients/lsp-trunk.el: prevent error when using lsp on non-file buffer

Protect against error when (file-name-directory filename) is nil which occurs on non-file buffer. We
were hitting this when using lsp-mode with the p4.el package which when diff'ing pulls files from
Perforce into non-file buffers and then activates a major mode that runs lsp.  This ends up calling
lsp-trunk-check-for-init.

* Update lsp-trunk.el

Co-authored-by: Sebastiaan Speck <12570668+sebastiaanspeck@users.noreply.github.com>

---------

Co-authored-by: John Ciolfi <ciolfi@.mathworks.com>
Co-authored-by: Jen-Chieh Shen <jcs090218@gmail.com>
Co-authored-by: Sebastiaan Speck <12570668+sebastiaanspeck@users.noreply.github.com>